add sell pint button

+++ src/app/new-keg/new-keg.component.ts
@@ -10,7 +10,7 @@ export class NewKegComponent{
   @Output() addKeg = new EventEmitter();
 
   submitForm(name: string, brand: string, price: string, alcoholContent: string){
-    let newKeg: Keg = new Keg(name, brand, parseInt(price), parseInt(alcoholContent));
+    let newKeg: Keg = new Keg(name, brand, parseFloat(price), parseFloat(alcoholContent));
     this.addKeg.emit(newKeg);
   }
 }
